Create new presenter to format a date
https://eaflood.atlassian.net/browse/WATER-4099

The update to the `health/info` page in PR https://github.com/DEFRA/water-abstraction-system/pull/376 requires a date to be formatted as a `DD MMMM YYYY at HH:MM:SS` string.

We have not converted a date to this format before so a new presenter is being created to ensure consistency should this format be required again.