Return `user` from `FetchUserRoles...` service
We decided that instead of simplying returning a boolean to say if the user was found, we instead return the actual user object. This allows us to check that the user was found (as it will be `null` if not), and gives us the option of using user data later on (eg. email address)