Add missing CRM_V2 table constraints (#404)
https://eaflood.atlassian.net/browse/WATER-4092

We are migrating the change billing account address functionality from the legacy code to this project. In our test DB, we previously added migrations to recreate the CRM_V2 tables `addresses`, `companies` and `invoice_account_addresses`. But to keep things simple we usually don't bother with constraints as unless we really need them, we prefer to avoid using them.

Well, in the [new Billing Account change address service](https://github.com/DEFRA/water-abstraction-system/pull/400) we're adding we do! 🤦 We depend on the constraints to raise [ON CONFLICT errors](https://www.postgresql.org/docs/current/sql-insert.html#SQL-ON-CONFLICT) for our logic to work.

So, in order to test our new service we need to add the missing constraints to our migrations. We're doing that change separately because [Add new Billing Account change address service](https://github.com/DEFRA/water-abstraction-system/pull/400) is already pretty big 😬.

---

When we did apply the constraints it highlighted we had a flaw in our model tests. The top-level `beforeEach()` should only have been cleaning the DB.

The model tests also had to be updated to deal with creating 2 invoice address records using the same default data; we can't have multiple records with the same start date and invoice account ID.