Add missing information to `/health/info` page (#376)
https://eaflood.atlassian.net/browse/WATER-4099

Finish off the work previously started on the `/health/info` page. The data that is missing that needs to be added in this PR is:
- Confirm Redis is running
- Display the latest results for the import app

The changes to add the Redis info have been moved into a separate PR https://github.com/DEFRA/water-abstraction-system/pull/402 so that this PR is just focusing on the import information.