Move auth logic into separate `AuthenticationService`
Unit testing our plugin logic was proving difficult as we need to inject a valid cookie into our request in order to hit the `validate` handler. However there doesn't seem to be any straightforward way of generating a cookie so we were looking at hardcoding it. Instead of this, we move the plugin logic into a separate `AuthenticationService`, which we can easily unit test.