Clean up response & demonstrate null args in test
Feedback in the PR suggested we have an explicit test that covers the service handling a null agent company and contact. After implementing the test we realised the response from the service was 'dirty' and confusing. If you passed in a null `agentCompany` you would expect to get a null `agentCompany` out.

So, we also took this opportunity to clean up the response from the service.