Add return schema lines and versions models (#408)
https://eaflood.atlassian.net/browse/WATER-4098

The next step in the two-part tariff work is to calculate the abstraction volumes for each return matched to a charge reference. After some investigation, we have determined that the data is held in `returns.lines`. The only problem is that the table is linked to `returns.returns` via `returns.versions`.

So, before we can start calculating volumes we need to add these 2 models to the project.