Add service to calculate total volume for a return
https://eaflood.atlassian.net/browse/WATER-4098

The next step in our two-part tariff journey is to calculate the total volume of water abstracted for a matched return. We know that a 'return' can have multiple 'versions'. So, the service first needs to identify the current version and find the matching `returns.lines` for it.

We know that not all lines should be included in the calculation. We'll need to know what volume was abstracted inside and outside the abstraction period. But we'll look at that in a subsequent change.

But we will have to handle returns that have no submissions in this change, plus include the result in the output for the `/check/two-part` endpoint.