Fix `/health/info` Redis connecion
https://eaflood.atlassian.net/browse/WATER-4099

After adding the missing information to the `/health/info` page in PR https://github.com/DEFRA/water-abstraction-system/pull/376 It was discovered that whilst it worked fine locally, when viewed on the `dev` server Redis was unable to connect. This was down the Redis connection on the AWS servers needing TLS to be enabled.

This PR will enable TLS by default for the Redis connection & add an environment variable to enable TLS to be disabled locally.

Whilst debugging this issue it was also found that if a connection to Redis could not be established the page would error rather than just display a connection failure message for Redis and continue to show the status of the other services. This behaviour was not present locally. After much hacking a fix has not be found for this issue. Given we have more urgent work currently this issue will be added to our issues backlog to investigate further at a later date.