Handle unable to connect
Found this post <https://stackoverflow.com/a/61133867> which put me onto some additional properties we can pass in the config that make the client throw an error as expected when a command fails to complete rather than using the under-the-hood Redis error handling.

This means should we fail to connect an error will be thrown which we can trap and display in our health info page.