Fix `/health/info` Redis connection (#411)
https://eaflood.atlassian.net/browse/WATER-4099

After adding the missing information to the `/health/info` page in PR https://github.com/DEFRA/water-abstraction-system/pull/376 It was discovered that whilst it worked fine locally, when viewed on the `dev` server Redis was unable to connect. This was down the Redis connection on the AWS servers needing TLS to be enabled.

This PR will enable TLS by default for the Redis connection & add an environment variable to enable TLS to be disabled locally.