Add service to calculate volumes for a return (#410)
https://eaflood.atlassian.net/browse/WATER-4098

> NOTE: we are still hacking and exploring this functionality. So, we are not writing unit tests until we have some confidence as to what exactly is needed.

The next step in our two-part tariff journey is to calculate the volumes of water abstracted for a matched return. We know that a 'return' can have multiple 'versions'. So, the service must first identify the current version and find the matching `returns.lines` for it.

We know that not all lines should be included in the calculation. We need to know what volume was abstracted inside and outside the abstraction period. To do this we need to work out the billable abstraction periods for the return.

We will also have to handle returns without submissions in this change.

Finally, we'll need to include the results in the output for the `/check/two-part` endpoint.

---------

Co-authored-by: Jason Claxton <jason.claxton@defra.gov.uk>