Revert ChargeVersion name change
We have just come back from 3 days of all the delivery team and its users referring to these as Charge Versions. In fact, their example formed the basis for an idea of doing the same thing for 'return versions'.

Clearly, that is how everyone refers to them, even if the UI page has the title 'Charge information'.

So, we revert that name change.