The Great Rename (#416)
https://github.com/DEFRA/water-abstraction-team/issues/97

Enough is enough!

Too many of our conversations when discussing this service follow the pattern of

> "The _charge version_, known as the _charge information_ in the UI, links to multiple _charge references_, which are _charge elements_ in the code. Each of these has multiple _charge purposes_, which are _charge elements_ in the UI. These records are used when generating a `bill run` which is a `billing batch` in the DB. A `bill run` is made up of `billing invoices` that are shown as `bills` in the UI".

This insanity is because things were either built before the content design had agreed on what they should be called, or they've changed over time but the code hasn't.

We're not going to go back and fix all the legacy code. But we have this confusion in the [water-abstraction-system](https://github.com/DEFRA) because we named our models to match the tables, which is the norm. We could rename the tables and our models to match the UI but then we'd break _everything_.

So, we are going to do what is within our control. We will rename our models and services to match how things are referred to in the UI (even though some of those names don't help!) Hopefully, this will reduce conversations like our example and help bring new team members on board. We'll still need to be aware of the name disconnect between the code and the DB. But at least our models will act as documentation for what is _actually_ what.

Anyone trying to match something from the UI to something in the code will have a much easier time if we rename everything.

**Renames**

- `BillingBatch` to `BillRun`
- `BillingInvoice` to `Bill`
- `BillingInvoiceLicence` to `BillLicence`
- `BillingTransaction` to `Transaction`
- `BillingChargeCategory` to `ChargeCategory`
- `ChargeElement` to `ChargeReference`
- `ChargePurpose` to `ChargeElement`
- `ChargeVersionWorkflow` to `Workflow`
- `PurposesUse` to `Purpose`