Tweak data tear-down
https://eaflood.atlassian.net/browse/WATER-4138
https://eaflood.atlassian.net/browse/WATER-4077

When running the [water-abstraction-acceptance-tests](https://github.com/DEFRA/) recently we have encountered an issue where tests fail because 2 'Big Farm Co Ltd 01' companies are returned when searching for an existing company record. One is a test record created the other is from a previous test.

Because the one created in a previous test is not flagged as `is_test` it remains even after running the tear-down. So, we need to tweak the tear-down process for the `crm_v2.companies` table. To make this distinction clearer, we also split the crm tear down into the separate schemas to match how the Db is actually structured.