Tweak data tear-down (#434)
https://eaflood.atlassian.net/browse/WATER-4138
https://eaflood.atlassian.net/browse/WATER-4077

When running the [water-abstraction-acceptance-tests](https://github.com/DEFRA/) recently, we encountered an issue where tests fail because 2 'Big Farm Co Ltd 01' companies are returned when searching for an existing company record. One is a test record, the other is from a previous test.

Because the one created in a previous test is not flagged as `is_test` it remains even after running the tear-down. So, we need to tweak the tear-down process for the `crm_v2.companies` table.