Update Knex config to avoid hanging promises
When trying to work with the DB in the tests we were getting

```
Jest did not exit one second after the test run has completed.

This usually means that there are asynchronous operations that weren't stopped in your tests. Consider running Jest with `--detectOpenHandles` to troubleshoot this issue.
```

Googling led us to [Having trouble running Jest, unhandled Promises](https://github.com/Vincit/objection.js/issues/1015) and [Need to explicit destroy knex in order to stop the script](https://github.com/Vincit/objection.js/issues/534). The second issue included a comment that setting the pool min size to 0 and idleTimeoutMillis to 500 would solve the problem. We tried it and it worked!