Add some more Jest args for debugging
The `--silent` false just means `console.log()` output will now appear which helps when debugging unit tests.

Also on our journey we saw

```
Jest did not exit one second after the test run has completed.

This usually means that there are asynchronous operations that weren't stopped in your tests. Consider running Jest with `--detectOpenHandles` to troubleshoot this issue.
```

So, we're doing as it suggests for now in case we see any others!