Add new bill run volume model (#440)
https://eaflood.atlassian.net/browse/WATER-4152

We are currently working on the SROC two-part tariff bill. The first step of the process in the UI is to review how the returns have been matched. Our superstar design team are working on an improved design for the returns review screen but to do that they need data.

Enter our mock data service!

We previously [Create endpoint to generate fake data](https://github.com/DEFRA/water-abstraction-system/pull/347) which was focused on a bill run. The design team would love it if we could do the same for the data the current returns review page uses.

The page relies on data in the `water.billing_volumes` table. So, the first step in generating the mock data is to add that model.