Fix issue allocating return volume
https://eaflood.atlassian.net/browse/WATER-4127

A issue has been spotted in the PR which deals with allocating the Returns volumes to the Charge Elements which was merged in https://github.com/DEFRA/water-abstraction-system/pull/436

The issue is that if the return volume exactly matches the `authorisedAnnualQuantity` the volume isn't allocated to that Charge Elelment. This PR will fix that bug.