Change where the allocated Return volume is held
https://eaflood.atlassian.net/browse/WATER-4127

In a previous PR https://github.com/DEFRA/water-abstraction-system/pull/436 which dealt with allocating the Return volumes to Charge Elelemts. It was assumed that the `billableAnnualQuantity` data item could be used to hold the Return volumes allocated to a Charge Element. However, in a very few cases this could already contain data which will cause the Return volume allocation process to not work properly.

Therefore this PR will create a new data item on the Charge Element called `allocatedReturnVolume` that will be used to hold any Return volumes allocated to a particular Charge Element.