Make test data helpers more 'unique' (#452)
https://eaflood.atlassian.net/browse/WATER-4144

We've been investigating changing our unit test framework recently (see [Investigate switching to jest](https://github.com/DEFRA/water-abstraction-system/pull/430) and [water-abstraction-ava](https://github.com/DEFRA/water-abstraction-ava)).

A key issue we'll have if we switch is that most other frameworks run tests asynchronously. This means our pattern of clean-setup-assert won't work. At best we can clean before any tests run but then the test data we set up must not clash.

To achieve that we need to make the test data the helpers generate as unique as possible, which means moving away from hard-coded values.

We're still playing with frameworks but there is nothing stopping us from making this change now ahead of a decision.

So, this change covers making the helpers generate more unique and fixing any tests that break.

** Notes

- Add randomInteger() to new GeneralHelper

Where a helper is hard-coding a UUID for a field is going to be easy to make it unique; we just call our `GeneralLib.generateUUID()` function.

But many helpers also need to generate references, licence numbers being a good example. If we are going to make these unique we're going to need a means to generate random numbers. Often they need to be within a certain range, for example, 100 to 999 because the number must be a certain number of digits long.

So, we're adding a new `randomInteger()` function which can handle this requirement. Doing so forced us to make a decision as to where to put it. We decided to create a `GeneralHelper` module that can be used going forward for any shared functionality in the helpers.