Fix abstraction period matching
Realised too late that what we get back from determine abstraction period service is an array of periods, not a single value! 🤦 So, needed to update the code to account for that resulting in a quick hack to the `_periodsOverlap()`.

Whilst doing that spotted we had a glaring mistake in the old logic as well (startDate > startDate - 🤦🤦🤦)