Add 2pt check for single licence
Whilst we're testing the engine we tend to focus on a licence at a time. This means we either go searching through the output or we add a temporary clause to the main query.

It'd be a lot easier if we could just ask for a single licence to be matched and allocated!