Oops! Fix line allocating
When we had fully allocated a line for one charge element we were setting unallocated to 0. If the return matched another element we'd then check `if (!line.unallocated)` intending only to match those lines which as yet had not been touched. Problem is `line.unallocated` also returns false, so we were resetting unallocated and allocating the line again.

Doh!

`'myProp' in myObject` appears to be the consensus for checking if a property exists nowadays. However, we need to negate. We prefer the simple approach to `if (!('myProp' in myObject))`.