Ensure consistent results in returns
We're assigning 'test IDs' to our records, including the returns to allow us to validate the output to a spreadsheet of scenarios we are now maintaining. But we realised we were not ordering the returns when querying for them, which means on different runs the order, and therefore the test IDs we assign could be different.

So, we apply an order using the return period start date then the return requirement. This is the order we want to 'offer up' returns to elements in the matching anyway.