Move test id and unallocated line assignment
This came when we spotted that unmatched returns were not getting test IDs assigned to there lines. To sort this meant moving the test ID assignment into the prep returns method. After doing that it made sense to also move creating the `unallocated` property there as well which meant we could ditch our funky logic.