What the!? Revert test assertion changes
This is the failing test output

```
Summary of all failing tests
 FAIL  test/plugins/charging-module-token-cache.plugin.test.js
  ● Charging Module Token Cache plugin › When the first call returns an invalid token › returns a null token

    expect(received).toBe(expected) // Object.is equality

    Expected: "FIRST_TOKEN"
    Received: null

      70 |       const result = await server.methods.getChargingModuleToken()
      71 |
    > 72 |       expect(result.accessToken).toBe('FIRST_TOKEN')
         |                                  ^
      73 |     })
      74 |
      75 |     it('does not cache the token', async () => {

      at Object.toBe (test/plugins/charging-module-token-cache.plugin.test.js:72:34)

  ● Charging Module Token Cache plugin › When the first call returns an invalid token › does not cache the token

    expect(received).toEqual(expected) // deep equality

    Expected: "SECOND_TOKEN"
    Received: "VALID_TOKEN"

      77 |       const secondCall = await server.methods.getChargingModuleToken()
      78 |
    > 79 |       expect(secondCall.accessToken).toEqual('SECOND_TOKEN')
         |                                      ^
      80 |     })
      81 |   })
      82 | })

      at Object.toEqual (test/plugins/charging-module-token-cache.plugin.test.js:79:38)
```

It looks like to get the tests to pass the assertions were changed in 2 of the tests in the problematic `test/plugins/charging-module-token-cache.plugin.test.js` test file.

This is 'surprising', especially in light of the fact that the values being checked for clearly do not match the stubbed response in the `beforeEach()`. Reverting the test assertions back to what they were in the Hapi-lab version (using Jest syntax) gets the tests passing again.

We never, _ever_ change a test assertion in this repo's test suite unless we are sure the original assertion is wrong. This is changing tests just to make them pass. 🤨