Fix error output in test runs
Traced the issue down to Hapi is expecting an error to have been thrown in order to determine a 500 response is needed. It's not getting one so is throwing one of its own. This means the test still passes, but not for the reasons we expect.

Ensure the mock resolves with an error though and Hapi is happy!