Fix tests for authorisation scoped routes
In the previous change we added config to say that a requests had to be authenticated to access `/bill-runs` and `/billing-accounts`. On top of that the user has to have a matching scope.

This broke our existing controller tests because they were not setup to include authentication in the test requests.

Thankfully Hapi allows you a way to bypass the authentication by including an `auth` object in the options you pass to [server.inject()](https://hapi.dev/api/?v=21.3.2#-await-serverinjectoptions).

To be honest, we wasted a lot of time trying to find examples of how to handle auth in tests many of which led us completely astray. They just don't seem to be out there. Thankfully, an old Stackoverflow post (https://stackoverflow.com/a/31955942/6117745) pointed us in the right direction.

It seems if you include an  `auth.credentials` object Hapi will just 'pretend' that was the result of the authentication strategy. This also means `validate()` and therefore `AuthService.go()` never get called (so no need to stub them).

But you still need to provide a matching scope for the route being tested to prevent a `403` being thrown.