Implemented but not working!
These are the changes we thought would get the tests passing but they don't! :-)

The reason is we'd like to log that an unauthorised request has been made, but ensure we return `404` as the response code. But we've overlooked we determine what status code to send back first, then whether to log anything based on that calculated code.

So, we're going to need to rethink things to achieve all the things we want.