Update the error plugin tests
Again we update our tests to reflect the outcome we want. Realised there was little point on creating routes with handlers that generate a scenario, if we then stub `ErrorPagesService` to return a specific result.

What `ErrorPagesService` is doing is very light weight. So, in this case we're happy to remove the stubbing altogether and let the scenarios (route setup in each test) dictate what the plugin needs to handle.