Enable and config authentication as default (#466)
https://eaflood.atlassian.net/browse/WATER-4085
https://eaflood.atlassian.net/browse/WATER-4155

As the final step of supporting pass-through authentication from the legacy UI app and before we can implement our first page, we need to enable authentication on all routes as the default.

This means in future, when we add a new route requests will be authenticated unless we disable authentication in the route's config.

The change involves telling happy to use auth by default then going through most of our routes and disabling it! 😁

Where it is applicable we update the route config to include the 'scope' authenticated users must have to be _authorised_ to access it.

We'll also use the opportunity to do a little tidying up of the route files as we go.