Implement footer links
This is a straight copy from the legacy UI. If using them locally bear in mind all system pages are expected to be rendered via the existing UI. So, go to <http://localhost:8008/system/health/info> and the links will work. Go to <http://localhost:8013/health/info> and they won't.