Links added
This is a working version of the links. The only issue is that they are always present whether a user is logged in or not.

To solve that problem we'll need to somehow get auth details passed into the 'context' used when compiling a view with [Vision](https://hapi.dev/module/vision/) and the [Nunjucks engine](https://mozilla.github.io/nunjucks/).