Add auth & user details to view context by default
https://eaflood.atlassian.net/browse/WATER-4155

When working on [Update page header with user links](https://github.com/DEFRA/water-abstraction-system/pull/474) we hit a blocker.

The header needs to display links but only if a user is authenticated (logged in). So, we need to pass this information into the 'context' used when compiling (rendering) a view.

Commonly, what you will see in controllers that return views is something like this

```javascript
  // NOTE: naming the object which contains the data collated by the controller `pageData` is a convention we use
  return h.view('info.njk', {
    pageTitle: 'Info',
    ...pageData
  })
```

The first arg is the name of the view to nunjucks template to render, the second is the 'context' containing the data the template will use.

We could have just done the following.

```javascript
  const pageData = { authenticated: request.auth.isAuthenticated, user: request.auth.credentials?.user }

  return h.view('info.njk', {
    pageTitle: 'Info',
    ...pageData
  })
```

If we did that though we'd need to do it in _every_ controller that returns a view. A better solution is to include it in the global context which is always passed in via the engine.

```javascript
  // app/plugins/views.plugin.js
  plugin: require('@hapi/vision'),
  options: {
    engines: {
      // The 'engine' is the file extension this applies to; in this case, .njk
      njk: {
        compile: (src, options) => {
          const template = nunjucks.compile(src, options.environment)
          return context => template.render(context)
        },
  // ...
```

Prior to this change the context is a static object also found in `app/plugins/views.plugin.js`.

```javascript
  context: {
    appVersion: pkg.version,
    assetPath: '/assets',
    serviceName: SERVICE_NAME,
    pageTitle: `${SERVICE_NAME} - GOV.UK`
  }
```

This change replaces that with a function that that will generate the global context including auth details found in the request each time a view is rendered. Whilst we're at it we'll update the plugin to match our conventions rather than being a straight copy of the **Vision's** example.